backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.0
slowapi>=0.1.9
python-json-logger>=3.0
prometheus-fastapi-instrumentator>=7.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
opentelemetry-exporter-otlp>=1.27
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
aiosqlite>=0.20.0
alembic>=1.14.0
